+++ janitor/functions.py
@@ -2,7 +2,6 @@
 
 import datetime as dt
 import inspect
-import itertools
 import re
 import unicodedata
 import warnings
@@ -4526,12 +4525,53 @@ def complete(
     fill_value: Optional[Dict] = None,
 ) -> pd.DataFrame:
     """
-    This function shows all possible combinations in a dataframe, including
-    the missing values.
+    This function turns implicit missing values into explicit missing values.
 
-    This function is similar to tidyr's `complete` function.
+    It is modeled after tidyr's `complete` function, and is a wrapper around
+    `pd.DataFrame.merge` and `pd.DataFrame.fillna`.
 
-    Individual combinations or combinations with groupings are possible.
+    Combinations of column names or a list/tuple of column names, or even a
+    dictionary of column names and new values are possible.
+
+
+    `Source <https://tidyr.tidyverse.org/reference/complete.html#examples>`_
+
+    .. code-block:: python
+
+            group	item_id	    item_name	value1	value2
+        0	1	    1	        a	1	4
+        1	2	    2	        b	2	5
+        2	1	    2	        b	3	6
+
+    To find all the unique combinations of `group`, `item_id`, and `item_name`,
+    including combinations not present in the data, each variable should be
+    passed in a list to the `columns` parameter::
+
+        df.complete(columns = ['group', 'item_id', 'item_name'])
+
+              group	item_id	    item_name	value1	value2
+        0	1	    1	        a	1.0	4.0
+        1	1	    1	        b	NaN	NaN
+        2	1	    2	        a	NaN	NaN
+        3	1	    2	        b	3.0	6.0
+        4	2	    1	        a	NaN	NaN
+        5	2	    1	        b	NaN	NaN
+        6	2	    2	        a	NaN	NaN
+        7	2	    2	        b	2.0	5.0
+
+    To expose just the missing values based only on the existing data,
+    `item_id` and `item_name` can be wrapped in a tuple, while `group`
+    is passed in as a separate variable::
+
+        df.complete(columns = ["group", ("item_id", "item_name")])
+
+            group	item_id	    item_name	value1	   value2
+        0	1	    1	        a	  1.0	    4.0
+        1	1	    2	        b	  3.0	    6.0
+        2	2	    1	        a	  NaN 	    NaN
+        3	2	    2	        b	  2.0	    5.0
+
+    Let's look at another example:
 
     `Source Data <http://imachordata.com/2016/02/05/you-complete-me/>`_
 
@@ -4563,7 +4603,7 @@ def complete(
     The null value can be replaced with the fill_value argument::
 
         df.complete(columns = ['Year', 'Taxon'],
-                    fill_value={"Abundance":0})
+                    fill_value={"Abundance" : 0})
 
            Year      Taxon     Abundance
         0  1999     Agarum         1.0
@@ -4577,10 +4617,10 @@ def complete(
     1999 to 2004? Easy - simply pass a dictionary pairing the column name
     with the new values::
 
-        df.complete(columns = [{"Year": range(df.Year.min(),
-                                              df.Year.max() + 1)},
-                                       "Taxon"],
-                    fill_value={"Abundance":0})
+        df.complete(columns = [{"Year": lambda x : range(x.Year.min(),
+                                                         x.Year.max() + 1)},
+                                "Taxon"],
+                    fill_value={"Abundance" : 0})
 
             Year      Taxon     Abundance
         0   1999     Agarum         1.0
@@ -4592,9 +4632,11 @@ def complete(
         6   2002     Agarum         0.0
         7   2002    Saccharina      0.0
         8   2003     Agarum         0.0
-        9  2003     Saccharina      0.0
+        9   2003     Saccharina     0.0
         10  2004     Agarum         8.0
         11  2004    Saccharina      2.0
+
+    .. note:: MultiIndex columns are not supported.
 
     Functional usage syntax:
 
@@ -4637,7 +4679,6 @@ def complete(
     :param fill_value: Dictionary pairing the columns with the null
         replacement value.
     :returns: A pandas dataframe with modified column(s).
-    :raises ValueError: if `columns` is empty.
     :raises TypeError: if `columns` is not a list.
     :raises TypeError: if `fill_value` is not a dictionary.
     :raises ValueError: if entry in `columns` is not a